requirements.txt
Pillow
requests
ast
pytube
moviepy
struct
discord
pandas
numpy
logging
subprocess

